make page-title-setup initializer fastboot compatible
Fixes #69

https://gist.github.com/kratiahuja/d22de0fb1660cf0ef58f07a6bcbf1a1c#addon-contains-browser-based-initializers-as-appinstance-initializersbrowserjs
